Rename mpi header
