+++ src/Out.cpp
@@ -18,7 +18,7 @@
 
 #include "Out.h"
 #include "DataStructs.h"
-#include "Mpi_structs.hpp"
+#include "MpiStructs.h"
 
 void Out::Progress(const Simdat& sim, const int itert) {
 	static int prog_print_last = 0;
